now move all sbc code out thru a staging folder
git-svn-id: svn://orca.physics.unc.edu/Orca@2673 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/IPE/IpeV4 FLT/ORIpeV4FLTModel.h
@@ -208,10 +208,10 @@
 - (BOOL) hitRateEnabled:(unsigned short) aChan;
 - (void) setHitRateEnabled:(unsigned short) aChan withValue:(BOOL) aState;
 
-- (unsigned short)threshold:(unsigned short) aChan;
+- (unsigned long)threshold:(unsigned short) aChan;
 - (unsigned short)gain:(unsigned short) aChan;
 - (BOOL) triggerEnabled:(unsigned short) aChan;
-- (void) setThreshold:(unsigned short) aChan withValue:(unsigned short) aThreshold;
+- (void) setThreshold:(unsigned short) aChan withValue:(unsigned long) aThreshold;
 - (void) setGain:(unsigned short) aChan withValue:(unsigned short) aGain;
 - (void) setTriggerEnabled:(unsigned short) aChan withValue:(BOOL) aState;
